+++ tools/lib/glbToB3dm.js
@@ -1,7 +1,8 @@
 'use strict';
-
 var Cesium = require('cesium');
-var defaultValue = Cesium.defaultValue;
+var getBufferPadded = require('./getBufferPadded');
+var getJsonBufferPadded = require('./getJsonBufferPadded');
+
 var defined = Cesium.defined;
 var DeveloperError = Cesium.DeveloperError;
 
@@ -11,25 +12,32 @@ module.exports = glbToB3dm;
  * Generates a new Buffer representing a b3dm asset.
  *
  * @param {Buffer} glbBuffer A buffer containing a binary glTF asset.
- * @param {Buffer} [batchTableJSONBuffer] A buffer containing the batch table to use for the b3dm asset.
- * @param {Buffer} [batchTableBinaryBuffer] A buffer containing the accompanying binary batch table section for the b3dm asset.
- * @param {Number} [batchLength] The number of features in the batch declared in the header.
+ * @param {Object} [featureTableJson] The feature table JSON.
+ * @param {Buffer} [featureTableBinary] The feature table binary.
+ * @param {Object} [batchTableJson] The batch table JSON.
+ * @param {Buffer} [batchTableBinary] The batch table binary.
  * @returns {Buffer} Buffer representing the b3dm asset.
  */
-function glbToB3dm(glbBuffer, batchTableJSONBuffer, batchTableBinaryBuffer, batchLength) {
+function glbToB3dm(glbBuffer, featureTableJson, featureTableBinary, batchTableJson, batchTableBinary) {
     if (!defined(glbBuffer)) {
         throw new DeveloperError('glbBuffer is not defined.');
     }
-    batchTableJSONBuffer = defaultValue(batchTableJSONBuffer, Buffer.alloc(0));
-    batchTableBinaryBuffer = defaultValue(batchTableBinaryBuffer, Buffer.alloc(0));
-    batchLength = defaultValue(batchLength, 0);
-    var byteLength = 24 + glbBuffer.length + batchTableJSONBuffer.length + batchTableBinaryBuffer.length;
-    var header = Buffer.alloc(24);
-    header.write('b3dm', 0); // magic
-    header.writeUInt32LE(1, 4); // version
-    header.writeUInt32LE(byteLength, 8); // byteLength - length of entire tile, including header, in bytes
-    header.writeUInt32LE(batchTableJSONBuffer.length, 12); // batchTableJSONByteLength - length of batch table JSON section in bytes. (0 for basic, no batches)
-    header.writeUInt32LE(batchTableBinaryBuffer.length, 16); // batchTableBinaryByteLength - length of batch table binary section in bytes. (0 for basic, no batches)
-    header.writeUInt32LE(batchLength, 20); // batchLength - number of models, also called features, in the batch
-    return Buffer.concat([header, batchTableJSONBuffer, batchTableBinaryBuffer, glbBuffer]);
+
+    var headerByteLength = 28;
+    var featureTableJsonBuffer = getJsonBufferPadded(featureTableJson, headerByteLength);
+    var featureTableBinaryBuffer = getBufferPadded(featureTableBinary);
+    var batchTableJsonBuffer = getJsonBufferPadded(batchTableJson);
+    var batchTableBinaryBuffer = getBufferPadded(batchTableBinary);
+
+    var byteLength = headerByteLength + featureTableJsonBuffer.length + featureTableBinaryBuffer.length + batchTableJsonBuffer.length + batchTableBinaryBuffer.length + glbBuffer.length;
+    var header = Buffer.alloc(headerByteLength);
+    header.write('b3dm', 0);                                    // magic
+    header.writeUInt32LE(1, 4);                                 // version
+    header.writeUInt32LE(byteLength, 8);                        // byteLength - length of entire tile, including header, in bytes
+    header.writeUInt32LE(featureTableJsonBuffer.length, 12);    // featureTableJSONByteLength - length of feature table JSON section in bytes.
+    header.writeUInt32LE(featureTableBinaryBuffer.length, 16);  // featureTableBinaryByteLength - length of feature table binary section in bytes.
+    header.writeUInt32LE(batchTableJsonBuffer.length, 20);      // batchTableJSONByteLength - length of batch table JSON section in bytes. (0 for basic, no batches)
+    header.writeUInt32LE(batchTableBinaryBuffer.length, 24);    // batchTableBinaryByteLength - length of batch table binary section in bytes. (0 for basic, no batches)
+
+    return Buffer.concat([header, featureTableJsonBuffer, featureTableBinaryBuffer, batchTableJsonBuffer, batchTableBinaryBuffer, glbBuffer]);
 }